refactor(workspace): cast_lossless, manual_assert, bool_to_int_with_if sweep
Auto-fix across 174 files:
- cast_lossless: replace `x as u32` with `u32::from(x)` for lossless casts
- manual_assert: convert if-then-panic to assert! / assert_eq!
- bool_to_int_with_if: replace if-bool-then-1-else-0 with u32::from(bool)
- match_bool: convert match on bool to if/else
- items_after_statements: move inner items before statements

Net -10 lines across protocol, browser, apply-patch, execpolicy, core,
tui, and 30+ other workspace crates.

+++ code-rs/tui/src/chatwidget/status_line_flow.rs
@@ -331,9 +331,9 @@ impl ChatWidget<'_> {
     fn status_line_context_remaining_percent(&self) -> Option<i64> {
         let context_window = self.config.model_context_window?;
         Some(
-            self.last_token_usage
+            i64::from(self.last_token_usage
                 .percent_of_context_window_remaining(context_window)
-                .clamp(0, 100) as i64,
+                .clamp(0, 100)),
         )
     }
